Downgrade paperclip gem to get rid of mimemagic gem
WARNING: This commit breaks the application.

mimemagic gem had proprietary code and was destroyed in 2021
This project uses paperclip gem to store attachments, my intention was to get rspec to work,
so I downgraded the gem instead of replacing it with a viable attachments gem like carrierwave.

WARNING: do not deploy to production